requirements.txt
streamlit
pandas
pymupdf
plotly
google-generativeai
python-dotenv
pillow
